requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
Pillow==10.4.0
cachetools==5.5.0
python-multipart==0.0.9